src/sagemaker_scripts/product_recommender/build_model/requirements.txt
# requirements.txt
numpy==1.18.1
jsonlines
pandas
google-cloud-bigquery
tensorflow==2.1.0
dill
